Using Errorable more widely now, so give it its own module
