+++ src/wizard.ts
@@ -1,24 +1,5 @@
 import { ShellResult } from "./shell";
-
-export interface Succeeded<T> {
-    readonly succeeded: true;
-    readonly result: T;
-}
-
-export interface Failed {
-    readonly succeeded: false;
-    readonly error: string[];
-}
-
-export type Errorable<T> = Succeeded<T> | Failed;
-
-export function succeeded<T>(e: Errorable<T>): e is Succeeded<T> {
-    return e.succeeded;
-}
-
-export function failed<T>(e: Errorable<T>): e is Failed {
-    return !e.succeeded;
-}
+import { Errorable } from './errorable';
 
 export interface ActionResult<T> {
     readonly actionDescription: string;
